#67: Rename protocol.py → provider.py and validator.py for clearer module names
