+++ src/promptum/providers/__init__.py
@@ -8,7 +8,7 @@ from promptum.providers.exceptions import (
 )
 from promptum.providers.metrics import Metrics
 from promptum.providers.openrouter import OpenRouterClient
-from promptum.providers.protocol import LLMProvider
+from promptum.providers.provider import LLMProvider
 from promptum.providers.retry import RetryConfig, RetryStrategy
 
 __all__ = [
