+++ src/promptum/providers/openrouter.py
@@ -12,7 +12,7 @@ from promptum.providers.exceptions import (
     ProviderTransientError,
 )
 from promptum.providers.metrics import Metrics
-from promptum.providers.protocol import LLMProvider
+from promptum.providers.provider import LLMProvider
 from promptum.providers.retry import RetryConfig, RetryStrategy
